Switched from 32bit offset/count read/writes to 64 bit for future proofing

+++ src/kernel/system/modules.c
@@ -65,15 +65,16 @@ uint32_t module_api(uint32_t func, ...){
         case MODULE_API_READ:
             vfile_t *file = va_arg(vars, vfile_t *);
             char *buffer = va_arg(vars, char *);
-            uint32_t offset = va_arg(vars, uint32_t);
-            uint32_t count = va_arg(vars, uint32_t);
+            uint64_t offset = va_arg(vars, uint64_t);
+            uint64_t count = va_arg(vars, uint64_t);
+            // printf("|%x %x %x|\n", (uint32_t)offset, (uint32_t)(offset >> 32), count);
             return_value = fread(file, buffer, offset, count);
             break;
             case MODULE_API_WRITE:
             file = va_arg(vars, vfile_t *);
             buffer = va_arg(vars, char *);
-            offset = va_arg(vars, uint32_t);
-            count = va_arg(vars, uint32_t);
+            offset = va_arg(vars, uint64_t);
+            count = va_arg(vars, uint64_t);
             return_value = fwrite(file, buffer, offset, count);
             break;
         case MODULE_API_CREAT:
